THIRDEYE-250 StarTreeRecordStoreFactoryFixedImpl should be more robust to file-system changes
This also adds logic to the collectionConsistency health check that ensures the
min / max time is the same as that in the index.

Further work should be done to ensure that every time segment is loaded in the
metric store, but this is just a small sanity check that will help us debug
production issues.

RB=448302
R=kgopalak,npawar,ajaspal
A=kgopalak