THIRDEYE-283 A user should be able to specify relative times in URI
The new URI route is:
"/dashboard/{collection}/{metricFunction}/{metricViewType}/{dimensionViewType}/{baselineOffsetMillis}"

This will find the latest available data time, then set the baseline to
baselineOffsetMillis before that.

This way, people can statically share a link and refresh it to get the latest
w/w dashboard, for example.

RB=497107
R=kgopalak,npawar
A=npawar