[PINOT-2262] Set mmap handles to null after the file is unmapped.
This moves us one step easier to debug the issue. We will get an NPE instead of
a JVM crash (which is what is happening now)

Fix usages of CustomBitSet to close it, and remove the finalize method from it.
The finalize method was used only in tests

RB=596782
BUG=PINOT-2262
G=pinot-dev-reviewers
R=kgopalak,jfim
A=kgopalak,jfim