[PINOT-2317] Fix connection validation logic on live instance change
Earlier logic to validate connections was validating one connection.
Added a validate method to a pool that can be invoked when liveinstance
changes, so that all idle connections can be validated.

Removed the part that waited indefinitely for a conneciton when handling
liveinstance change (not applicable in the new code)

Updated SingleRequestHandler to respect timeout, as well as limit the number
of retries (to 3) when we cannot get a valid connection to the server

Added a unit test

Refactored the tests to use common class for server

RB=606801
BUG=PINOT-2317
G=pinot-dev-reviewers
R=kgopalak,jfim,atumbde,mshrivas
A=kgopalak