[PINOT-2398] fix for logging an exception in broker
We are logging an exception when a server has an exception processing a query,
and sends a DataTable that has no dictioary or schema, but just metadata. This
is not an exception, so handle the case gracefully.

Also, put in a check for version, a start for future-proofing.

RB=617302
BUG=PINOT-2398
R=kgopalak,jfim,atumbde,mshrivas
A=kgopalak