Cleanup logging in our unit tests
Our tests spit out reams of output to stdout generating multi-MB logs
and makes it harder to debug. Most of these statements seem to be debug
printfs written during tests and are not useful for automated testing.
This commit turns most of those into trace output so that we can enable
those if required but we should mostly not see those. This is only part-I
and is not a complete cleanup.

RB=617627
G=pinot-dev-reviewers
R=kgopalak,jfim,ssubrama,mshrivas
A=ssubrama