[PINOT-2381] Duplicate bytebuffer before bulk-copying data
FixedByteSingleValueMultiColReader supports bulk copy method to read
data. The ByteBuffer usage in that is not safe. This change duplicates
bytebuffer before copying it. From indications on web, even absolute get
and duplicate methods are not guaranteed to be thread-safe. ThreadLocalBuffer
gurantees absolute safety at the cost of performance. Plan is to test
the performance before introducing that change.

RB=622093
G=pinot-dev-reviewers
R=kgopalak,jfim,ssubrama,mshrivas
A=kgopalak