[PINOT-2645] Performance fixes for operator flattening and AndDocdid iterartor
Flattened the query tree so that re-ordering the operators works well with subtrees
We bring up a node's children one level if the node and its parent has the same AND operator
(or OR operator).

Further, on a filter like
   (a AND (b AND c))
where we have an index for the 'b' operand, then the fastIterator
constructed a full set of documents to evaluate the predicate against
'a', another full document list.In this case,
we revert to what the slowIterator does-- create the AndDocIdIterator.

RB=654014
BUG=PINOT-2645
G=pinot-dev-reviewers
R=kgopalak,jfim,atumbde,mshrivas
A=kgopalak