[PINOT-3101] Fix connection leaks on server restarts and connection timeouts
In ScatterGatherImpl.SingleRequestHandler, we need to cancel an outstanding future if we timed out getting a
connection. Otherwise, the counts in AsyncPoolImpl will go wrong, and we will be waiting on connections to be
released when none will be.

On a graceful close from the server, we need to destroy the resource associated with that connection (from
the point of view of AsyncPoolImpl)

Adding more logs to debug connection issues

RB=735254
BUG=PINOT-3101
G=pinot-dev-reviewers
R=kgopalak,jfim,atumbde,mshrivas
A=kgopalak,atumbde