[PINOT-3151]: Fix incorrect min/max value in metadata for star tree segment.
For segment creation, we use StatsCollector that makes a pass over the
documents to collect stats that are later used to build the segment.
For star-tree, we append the aggregated docs at the end of raw docs, and
stats collection happens on them as well. This RB addresses the issues
related to that, and ensures that only appropriate stats collection
happens for aggregated docs:
1. Unique values are collected, so dictionaries created appropriately.
2. Sorted checking is done, as aggregated docs can potentially impact
that.
3. Max number of multi-values are collected for multi-valued columns.
4. Min/Max values are 'not' updated with aggregated docs.

RB=737616
G=pinot-dev-reviewers
R=kgopalak,jfim,atumbde,ssubrama
A=xajiang