Removed un-necessary recursive calls to 'createAggDocForAllNodes'.
The code to create aggregatedDoc for all nodes in the star tree needs to
skip its recursive calls for star nodes. Not doing is so is not only
waste of CPU resource, but also may potentially cause NPE when the only
child of a star node is star child (ie no other nodes).