StarTree: Address conflict between split-order and skip-materialization config specs.
The split-order and skip-materialization are mutually values, ie a
dimension cannot be specified in both.
In the current code, if skip-materialization is not defined, we compute
a default value for it. And in case a split-order was explicitly
defined, but skip-materialization was not, we still give preference to
the default skip-materialization value, and remove the dimensions
appearing in both lists from split-order. This is not correct.

With this PR, we give prefernce to split-order if it was explicitly
defined by the user.

Testing: All star-tree tests + manual testing.