[PINOT-3198] Part 1 of 2 commits to change SegmentOnlineOfflineStateModel to contain the new CONSUMING state
We need to change the state model in order to introduce the CONSUMING state for low-level kafka consumers.
Since the new states and transitions are a super-set of the existing states and transitions, a straight-forward
way to do this is to upgrade the servers so that they process the new transitions (but they do not receive any yet),
and then upgrade the controller with code that updates the state model definition in place. The change to the
controller should be done after all the servers already have this change.

Tested as follows:
- Bring up a local cluster, add a table and a segment to the table. Ensure it is ONLINE and queries work.
- Shut down the server, apply this change, and start up server again. Add another segment and ensure that
  it is online and queries work.
- Shut down the controller, apply the next change (Part 2 of 2), validate that the state model has been updated.
- Add another segment to the table, and ensure that it is ONLINE and queries work.