[PINOT-3198] Part 2 of 2 commits to change SegmentOnlineOfflineStateModel to contain the new CONSUMING state
Added a configurable option for the controller to change the state model to prepare the cluster for
adding low-level kafka consumers.

Any new clusters created will be created with the new version of the state model, but the servers will
not go through the new states since the controller only updates the IDEALSTATE to contain ONLINE or OFFLINE
states.

Tests:
- Ran integration tests with a version of the server that does NOT have the changes for the new state transitions
  (the clusters in these tests were created with the new state model).
- Manually tested the case where a controller upgrade for a previously created cluster leaves the statemodel
  untouched
- Manually tested that adding the config to turn on the statemodel update works, and updates the state model
  as expected, and we can add new segments and queries work after the update.