[PINOT-3204] Plumbing the path for adding a new table.
Changed kafka consumer type in config to be a list rather than a single type

Added a new config variable for num replicas of low-level consumer. There is some discussion on this,
so I may change this later.

Added a class for segment metadata for LLC segments.

Added methods to create initial segment metadata for low level segments