Optimizing performance of splitting on time column.
In star trees with large number of leaf nodes, splitting on time columns
was creating/flushing byte buffers per leaf node, causing huge
performance issue. In this PR, a new version of the StarTreeDataTable is
introduced that will minimize the byte buffers created/flushed.

Merging the two versions of StarTreeDataTable is somewhat non-trivial,
and will be merged in the upcoming PR's. This is so we are unblocked
until then.