Disabling retention in integration tests. (#290)
During a recent debugging of integration test failure, it was found that
we set a retention value of 900 days during setting up tables for
integration tests. Most these tests use the airline data for which some
records are now > 900 days old, and getting deleted during the test.

The fix is to disable retention by not setting valid retention time
value and unit.