[PINOT-3200] Adding capability in broker to select HLC or LLC segments for routing based on a selector class
In order to migrate live use cases to use low-level kafka consumers, we need to have them generate both high-level
as well as low-level kafka segments for sometime, and have the brokers route some percentage of requests to low-level
segments. We increase the percentage over time, until we route only to low-level segments, and then drop
the high-level kafka segments later.

This PR addresses the capability in the broker to be able to selectively route requests according to a
specified selector class (configurable).

The feature of being able to specify a RoutingTable class has been removed to reduce the complexity of handling
low-level segment routing while still having a custom class for high-level and other segments.