[PINOT-3349]: Adding new interfaces for StarTree & StarTreeIndexNode.
This is a pre-cursor to the re-implementation in-memory & on-disk
representation of star tree:
1. Two new interfaces StarTreeInterf and StarTreeIndexNodeInterf have
been added.

2. All references to StarTree and StarTreeIndexNode have been replaced
to refer to the the new interfaces instead. The only exceptions are
clients that explicitly work on a specific version of StarTree (eg
OffHeapStarTreeBuilder that builds the current version of star tree).

3. Added a util class StarTreeSerDe that will handle the
serialization/de-serialization of the existing as well as future
implementation of the StarTree representation.

4. Validated by running all star tree tests.