[PINOT-3479] Provide best effort segment size if the segment is being pre-processed
Table size reader may find that the segment on disk is not yet pre-processed. So,
expected sub-directory v3/ may be missing. This caused size reader to throw an unhandled
exception. The server will not provide the best effort size estimation for this case.
If the expected v3/ directory is missing, server will provide aggregate length of files
from parent directory

Testing: UT + manual