Added QueryContext to pass information about the query to lower layers
Added query context class that can be populated with more detailed
information about the requested query to lower levels. This is
currently populated only with the IP address of the broker to start with.
In future, this IP address will be replaced with the broker id sent as
part of query request to the server.

Testing: Manual testing by starting services and issuing query