Added QueryContext to pass information about the query to lower layers (#480)
* Added QueryContext to pass information about the query to lower layers

Added query context class that can be populated with more detailed
information about the requested query to lower levels. This is
currently populated only with the IP address of the broker to start with.
In future, this IP address will be replaced with the broker id sent as
part of query request to the server.

Testing: Manual testing by starting services and issuing query

* Updated broker-server protocol to pass broker id as part of instance
reqeust to server.

Added broker id as part of request to the server. This broker id is
added to the query context in the server. The field is optional.
If the field is missing, server will log the ip address of the
broker making the request.

This protocol change is backwards-compatible.

Testing:
1. Test with new versions of broker and server
2. Test with old broker and new server
3. Test with new broker and old server