Re-architected server query execution to hand over query to the scheduler for execution. (#536)
This commit changes the server to pass query request to the
scheduler for execution instead of executing the query in
the receiving Netty server thread. Scheduler by default has two
executor pools - QueryRunners which will execute the query, and
QueryWorkers which are used by the query runtime to parallelize
operators across segments. Added a First Come First Served query
scheduler that matches existing scheduling functionality by
immediately sending the query to the QueryRunners for execution.
This commit also adds a ScheduledRequestHandler which _will_ replace
SimpleRequestHandler to use QueryScheduler in future.

Current code continues to use SimpleRequestHandler. The impacting change
of this commit is to run queries in separate executors and to return
ListenableFutures to the caller.

Testing:
1. Additional unit tests and existing integration tests
2. Query execution on local cluster
3. Performance testing is in-progress