Add routing options to BrokerRequest (#629)
Add debug options to BrokerRequest, in order to pass parameters to the
query routing phase (and eventually other parts). Added two options
"FORCE_HLC" and "FORCE_LLC" that respectively force the query routing to
send the query to the HLC or LLC counterpart of a realtime table and
error out if that part does not exist. Those options do not do anything
for offline tables.