Added a utility to reduce number of replicas for an offline table
If we want to reduce the number of replicas for an offline table, then
running helix AutoRebalanceStragy to recompute partitions does not work.
It does work if we want to increase number of replicas.

Added a separate utility command to reduce the number of replicas.

Also fixed the Rebalancer to account for IdealState changes during the time
th command is run.