Limit number of groups in Aggregation-GroupBy query results.
In AggregationGroupByOperator, we compute all groups at segment level.
This can cause high heap usage (or even OOM), if number of groups is
extremely large (multi-million groups range). This change limits the
number of groups returned by group by operator to a pre-defined size.

1. Added MAX_NUM_GROUPS_ALLOWED that defaults to 100K, and can be
   overwritten by unix env variable "pinot.max.aggr.groups".

2. ResultHolders switch from array mode to map mode, and start trimming
groups (based on sort order) to cap the number of groups to the limit.

3. Benchmark results, on 5.5B records, 3.6M unique keys (per segment), 17 segments:
   - Takes 800 MB as opposed to > 2 GB, with similar latency.
   - With ~40M keys per segment, existing code runs OOM, where as this
     change can still perform within < 1 GB of memory.

4. Added unit tests for the same.