Add a limit on number of aggregation groups in query response.
Queries with huge numbers of aggregation groups (millions range)
increase heap usage, and may lead to OOM. This change adds a
configurable limit on the number of aggregation groups, once query
execution crosses that number, it skips processing rest of the
documents, and returns truncated results.

1. Added a server config 'num.aggr.groups.limit', with a default value
of 1 million.

2. GroupByExecutor will skip processing all remaining blocks (with a
warning logged) once this limit is reached.