Remove call to join() when the consumer thread replaces realtime segment
The call to stop() can happen either from the consumer thread itself (when the segment
is committed or retained), or from the helix thread that executes the CONSUMING to ONLINE
state transtion. In the latter case, we want to call join() so that we wait for the
consumer thread to exit before examining the offsets. In the former case, we should
not be calling join().