Do not include recently pushed segments in calculating number of replicas (#692)
When a segment is pushed to the controller, it is first added to the ideal state and after a while it makes it to
the individual Pinot servers and have it in the external view. We do not want to include such segments in the
calculation for number of replicas. Otherwise during the period of segments being pushed, we will mistakenly
conclude that the table has fewer replicas than needed.