Add multiple or equality predicate optimization (#728)
Optimize query predicates where multiple equality and IN clauses are
linked by OR operators. For example, this rewrites the query predicate
a = 1 OR a = 2 into a IN (1, 2).

This optimization removes redundant OR operators ("pull-up") if the
OR node contains a single predicate. It removes duplicate values
across linked predicates (eg. a = 1 OR a = 1 is collapsed to a single
equality statement). Equality predicates of different types are
joined (eg. a = 1 OR a IN (2,3) becomes a IN (1,2,3)). Singular
predicates are turned into equality predicates (eg. a IN (1) OR
a = 1 becomes a = 1).