Add optional clause sanity checks to SELECT (#734)
Add sanity checks for optional clauses to select statements to ensure
that each clause occurs only once.

In a sane grammar, this is handled at the grammar level by specifying
a strict ordering of clauses as in standard SQL. SQL-92 and ulterior
releases specify the order as SELECT FROM WHERE GROUP BY HAVING (and
WINDOW for SQL-2003). LIMIT and TOP are non-standard extensions
inherited from MySQL, PostgreSQL and others, which normally happen at
the end (except for T-SQL's TOP clause which can happen as a prefix,
as in SELECT TOP 5 ...).

In PQL, optional clauses can be specified in any order, mostly for
legacy reasons (compatibility with SenseiDB's BQL). As such, this
check is done at a later stage of query compilation as opposed to
being a grammar-only change.