Remove spuriously emitted LLC routing metrics (#758)
When receiving a query, we check which tables exist in the routing
table by issuing a routing request to the routing table. To determine
whether or not a table has an offline part and a realtime part, we
check both the realtime and offline names for existence in the routing
table.

If a table does not have a routing table entry in the normal routing
table, we assume it is an LLC request and route it as an LLC query.
This unconditionally increments a counter for the number of queries
sent to LLC.

Those two behaviors put together mean that for every query to any
table, we increment the LLC query count, even if that table is
offline-only, has the "_OFFLINE" suffix or even if it does not exist.

This removes these spuriously emitted LLC metrics by adding a method
to check if a table has a non-empty entry in the routing table and
using it to determine if this is a single table query or a federated
query.