Remove trailing nulls ('\0') from string values during indexing. (#755)
NULL ('\0') is the default padding character for string values in
dictionary. If input strings contain trailing null characters, when
padded, different values can map to the same padded string in the
dictionary. This currently triggers an assert in the segment generator,
that in turns stop Kafka consumption in realtime nodes.

Stripping trailing nulls from incoming string values addresses the
issue. This also means that two values 'abc\0' and 'abc' will be
interpreted as the same value 'abc', and this is a behavior we are
choosing to keep.