Fix path name used for deleting expired segments
We are currently not deleting any segments from the controller store because we
look for the segment in the directory tableName_OFFLINE (or tableName_REALTIME)
whereas the segment is actually in the directory tableName (without the type
extension).

This change moves the segments to another directory under _localDiskDir, from which
segments can be deleted periodically as needed to reclaim space.