[PINOT-4062] Auto-create segments in CONSUMING state if we do not see one in a partition (#858)
* [PINOT-4062] Auto-create segments in CONSUMING state if we do not see one in a partition

This feature is useful to fix kafka consumer errors (e.g. offset not found). In this condition
the segments will still show up as being in CONSUMING state, except that the servers
have stopped consuming, and are waiting for some other server to commit the segment so that
they can get the right segment. If each server thinks they are the only one that is having
a problem, then consumption will stall.

We have a metric for number of bytes consumed, which should drop to zero.

In this case, the operator can just remove the segment in CONSUMING state for that partition,
and we will auto-create the segment on the next run of ValidationManager.

It may be useful to run this functionality more than once in an hour, TBD.

* Addressed review comments

* Resolved merge conflicts