Fix auto-creation of segments
Several fixes applied:

The auto-creation loop was skipping over segments if consecutive partitions had missing segments.

The auto-creation logic was always taking the earliest kafka offset (assuming the error is
that of offset not being found), but this may not be the case. So, if we have an older segment whose
end offset is higher than the earliest available kafka offset, we should pick the higher one.

Added tests for these cases.