Adding log messages during validation run
In a recent incident, we found that the validation manager did not correct a segment that was
marked OFFLINE instead of CONSUMING in idealstate. A local run of the validation manager
(pointing to the real cluster) seemed to run fine. The logs did not show exceptions.

With these logs we should be able to debug better should it happen again.