Log an error and update metric when zookeeper updates fail (#968)
* Log an error and update metric when zookeeper updates fail

We have retries for helix updates, but if an update fails even after multiple retries
(as has been seen sometimes), we need to know about it.

Changed the retry policy to a linear fixed-delay policy with max 10 retries, attempting
an update every 1/2 second. (earlier we had exponential back-off with 5 max attempts).
This places a maximum time delay of 5 seconds to update zookeeper. Hopefully, zk is not
busy for more than 5 seconds.

Should an update fail after all these retries, we bump a metric, that we can alert on.
If this alert files, some manual repair of zookeeper may be needed, production experience
will tell better.

* Update log messages to provide more information when an exception happens