Limit number of servers hit per query for LLC (#991)
Add code to limit the number of servers hit per query for LLC realtime tables,
similar to the one for large clusters. Refactor both the LLC routing table code
and the large cluster routing table code to share the routing table generation
code and only passing a different set of segments and replicas based on the
idiosyncracies of each table type (eg. for LLC, we don't want multiple
CONSUMING segments, for offline we only care about ONLINE, etc.).

Remove leftover debugging code that accumulated the number of replicas in a
fixed size array.