Cache totalDocs and totalRawDocs in SegmentMetadataImpl, instead of (#989)
fetching from config within getter.

In SegmentMetadataImpl, the getters for totalDocs/totalRawDocs fetch the
value from config (which is a hash-lookup). With large number of
segments, these fetches start showing up in the CPU profile. This change
is to fetch the values once in the constructor, and return it in the
getter.

This also required adding the missing override for getTotalRawDocs in
RealtimeSegmentImpl, as in real time, we don't know the current number
of docs consumed upfront.