Integration Test
Added an integration test that contains both delete APIs. Since the integration test setup had strange characters in the segment names, I also added an URL Decode in the Restlet so that the APIs will both have the correct file names that they expect.

The Delete API takes in a required table type, but the Get API to drop segments doesn't require a table type.