Change LLC segment name ordering
Change the LLC segment name ordering to take into account the
timestamp in segments instead of throwing an exception if segments
have the same partition and sequence number.

In certain situations, there can be segments with the same sequence
number present in the external view. For example, by disabling an
instance in Helix and then emptying a table, this will defer the
deletion of the segment on the disabled instance, so the previous
segment will be shown as offline in the external view. When the newly
created segments match with a sequence number from the previously
existing segments, this will cause an exception while building the LLC
routing table because segments are filtered after being added to the
sorted set of segments.

This fixes this issue by allowing the sorting of segments to proceed.