Fix brokers reporting STARTING for a long time (#1025)
Fix brokers reporting a service status of STARTING for a long time,
due to some partitions not being present in both external view and
ideal state.

Correct the check to only consider a broker as being STARTING if the
broker is part of the ideal state when the corresponding external view
partition does not exist. Add a check to ignore disabled resources. Add
a check to skip partitions that are assigned to OFFLINE status in the
ideal state.