Segment Deletion API  (#1029)
Originally started with PINOT-4349, to fix the following bug: 
* Previously, if the tableType was null, and the table has both realtime and offline components, we look for the segment in the realtime table, don't find anything there, and then we throw an error. Now, we check to ensure the table type is correct by looking at the segment metadata.

While we were fixing the above bug, we discovered:
* We have two APIs to delete segments, one using a curl DELETE call and one using a curl GET call. The GET API was using an incorrect codepath to delete, aka, the logic that we had in the GET API was never run; we were using the delete API codepath. Therefore, I fixed the codepath for the GET API so that both use the same underlying method but run through the right request.

We also decided to change the DELETE API to take a required tableType parameter so that clients don't accidentally delete all their segments. We validate that this tableType parameter is not null and is a valid tableType and throw back an IllegalArgumentException with a message to the client if it isn't. Previously, the DELETE API did not have this parameter. For the GET API, we left the tableType parameter as optional. I updated the swagger docs for clients accordingly. 

Added an integration test that contains both delete APIs(using the GET and the DELETE). Since the integration test setup had percentage and space characters in the segment names, we added an URL Encode in the Restlet so that the APIs will both have the correct file names that they expect. As a result of this, we discovered that on the GET/DELETE API side, Restlet doesn't take care of URL Decode from their side, so I added an URL Decode to both of those Restlets.

While we were writing the integration test, we also discovered that the segment list API has a bug where if it's called immediately after a segment is deleted, it will return an error. Essentially, since segment deletion is really slow(it takes 2 seconds), there could be a race condition in that time where the deleted segment is in the list we fetch, but then when we try to get the instance state map of the segment from the ideal state, it is null.

Notes:
* There is a bug in the segment deletion manager where it doesn't actually move files that end with .tar.gz, so there is no good way to tell if a segment has been "deleted." This is important because we repush all offline segments so that we can test both APIs, so we have a Thread.sleep in the integration test to account for this. When we fix this bug, we should revisit the integration test.
* We don't have tests for delete all realtime segments from either API because of HLC/LLC, but it is a TODO for when we finish migrating to LLC.
* Because of the above problem, I manually tested delete all realtime segments for both APIs and delete all segments(both offline and realtime) for deleting through the GET API.