Filter optimization to avoid unnecessary predicate evaluation. (#1027)
Replace AND/OR operators with EmptyFilterOperator if one child (all
children for OR) of AND Operator always evaluates to false. Doing so in
planning phase completely eliminates thread overheads during execution
phase. In an example use-case with 180 segments, where most segments are
eliminated due to time filters in query, this change provided > 25%
improvement in QPS, while improving latency by 10%.

1. All filter operators now implement alwaysFalse(). Leaf level filter
operators use predicate evaluator for the same. While non leaf filters
(AND/OR) decide that based on their children (one child is alwaysFalse
for AND, all children alwaysFalse for OR).

2. While creating FilterPlan, return an emtpy filter operator for all
AND's that have at least one child that always evaluates to false. This
helps optimize the filter tree by collapsing parts of the tree that are
alwaysFalse.

3. Removed set/getPredicate() from BaseFilterOperator, leaf level filter
operators take predicate in constructor now.

4. Added unit test to test the always false optimizations.