Cleanup: Remove fieldspec dependency from mutable dictionaries.
MutableDictionary can be reused for creating on-the-fly dictionary for
TransformBlock, where the data has been transformed, thus does not have
a dictionary of its own.

In the current code, MutableDictionary requires FieldSpec in the
constructor, but does not really use it, other than getting the column
name for the FieldSpec. Removing the FieldSpec dependency would allow it
to be reused for TransformBlock (which does not have a FieldSpec).