[PINOT-4446] Revert back to exponential back-off and introduce locks to serialize idealstate updates to same table (#1036)
Exponential back-off introduces a randomness so that retries have a higher chance of
succeeding, so changed the retry policy back to exponential back-off.

In addition, introduced a lock to serialize updates to idealstate on the same table.
This helps when there are multiple LLC partitions of the same table completing
around the same time. While indepdendent threads can process the state changes
during segment completion, the lock ensures that the updates happen in a serial
fashion.

There is still an undesriable behavior on failures, though. For any reason, if
the update to idealstate fails during segment completion, we will be left with the
scenario where the completing segment is in CONSUMING state, the PROPERTYSTORE
entry for the completing segment has already been marked as DONE, and the
PROPERTYSTORE entry for the next segment has already been created.

When this happens, the servers will attempt to re-complete the segment, and will be told
to either download or keep the segment. But they will never start a new segment.

TBD how to fix this.