Re-implement MutableDictionary with primitive type, when possible.
There are two motivitations behind this change, both of which will
benefit by reduced memory consumption by avoiding java objects where
primitives could be used.

1. Mutable dictionary implementations are currently used for realtime
segments in consuming state. This increase memory consumption (due to
non-primitive key/values) throughout the life of the segment in
consumption.

2. No-dictionary group key generator can now use the mutable dictionary
as a on the fly dictionary.

Added unit tests for the new implementations.