Auto-fix segment consumption if idealstate update fails during LLC segment completion (#1049)
* Auto-fix segment consumption if idealstate update fails during LLC segment completion

When a segment completes, we update zk with two changes:
1. Update proprtystore with a new LLC segment
2. Update idealstate to mark old segment in ONLINE and add a new segment in CONSUMING
These two updates cannot be done in a transaction. It is possible that the controller
fails after step 1 but before step 2. In that case, we had called the code to fix
things on a controller getting leadership.

It is also possible that the idealstate update fails due to concurrent updates to
the idealstate znode (e.g. retention manager, deletes by operator, etc.). In these
cases, we have a retry, but no guarantees.

With this change, we call the same logic that we call on leadership change.

Also changed the logic that detects and fixes the condition to use priority
queues to make the code more readable.

Existing unit tests cover this case already

* Addressed review comments

* Addressed rest of the review comments