[PINOT-4358] Added code to remove old LLC segments that are marked IN_PROGRESS but in OFFLINE state
If LLC consumption fails with a permanent failure (or after a certain number of retries), we mark the
segment to be in OFFLINE state (and let the auto-creation mechanism come by and create new segments).
However, we leave these OFFLINE segments as cruft, never to be removed.

Changed RetentionManager to delete these segments.

Added a test.