Adding longer timeout (#1098)
The DeleteAPIHybridClusterIntegrationTest has two test cases(deleteRealtimeSegmentFromDeleteAPI and deleteRealtimeSegmentFromGetAPI) that recently started failing after the following commit: 
https://github.com/linkedin/pinot/commit/8fa3d977ff700dd1c5db602d81bc164bac91cfeb 
These tests pass locally, but have stopped passing on Hudson, so my guess is that it needs a longer timeout. We will investigate the root cause, but meanwhile, to unblock the release, we will increase the timeout.