[PINOT-4603] Extended the protocol with the lease message
Added protocol extension on the controller side to field the build-extension time
to be sent by the server.

The SegmentCompletionManager indicates to the server in the COMMIT response the max
time that the server can take before the SCM declares it dead. The SCM expects the
server to send a (newly introduced) extendBuildTime() message that extends the time
as often as needed.

Added test cases