[PINOT-4637] Configure the average number of multi-values per row.
For multi-valued columns in realtime segments, we allocate memory initial memory assuming that the average
number of values in any row for that column is 10. Changed the default to 2, and made it configurable
for individual server instances if needed.

Modifed the tests to print the seed value in case of exceptions