[PINOT-4462] Step 1 to add auto-cleanup of deleted segments
We copy the deleted segments without preserving timestamp. This will allow us to
write a backkground thread that removes files from the DELETED directory that are
older by N or more days.

We need to let this code be in there for N days (if we want N day retention of
deleted segments)