Separate IndexLoadingConfig from TableDataManagerConfig
This PR is a prerequisite of RELOAD message.
When loading a segment, we should pick up the latest table config.
Old behavior: when we create the TableDataManager, we set the IndexLoadingConfig and it becomes immutable, which causes new segment unable to pick up the latest table config.
New behavior: create a new IndexLoadingConfig per segment loading so that segment loading config is in sync with latest table config.

Fixed a bug where SegmentMetadata.getIndexDir() return inconsistent result for v3 format when loading with different methods.
Added integration test to test triggering inverted index creation.