[PINOT-4475] Handle SimpleDateFormat correctly when creating segments (#1200)
* [PINOT-4475] Handle SimpleDateFormat correctly when creating segments

If time column is of SimpleDateFormat type, the we need to set the start and end time
correctly in the segment.

Input files to pinot segment creator (for pure offline use cases) may contain time column values
like 20170324 to indicate Mar 24 2017. If we take this value to be from epoch, we will get year 57194.

Modified SegmentGeneratorConfig to take additional configuration for time type. The default
is epoch, so that the default behavior remains unchanged. Setting the time format correctly
will now re-adjust the start/end time of the segment if needed, so that the start and end time
in the segment metadata is reflected correctly.

* Added unit test. Addressed review comments