Add support for partitioners in LLC realtime.
This PR adds support for partitioning functions to prune on-disk segments
in real time tables. While generating Kafka events, events are keyed on a column and partitioned using a specified algorithm.
This enables events with a specified key to go to a particular partition.
This algorithm is duplicated in Pinot so that we can verify that the keys are going to the correct partition.
If an entire segment has keys which satisfy a given algorithm, we write the
key information and the partition function in the segment metadata. If a query
comes to Pinot for a particular value of the key column, we can prune out the
segments which don't satisfy the algorithm for the given value.
We provide partitioners for some common partitioning functions.
This feature is tested using the LLCPartitioningClusterIntegrationTest.