Fix the NPE for realtime segment
In realtime consuming segment, we first add value to dictionary then update inverted index.
If we query a column with inverted index, it is possible that the inverted index has not been generated.
The fix is filter out all null inverted index.
This will not affect the accuracy of the results because we just ignore a single under-indexing record.