[PINOT-4726] Remove excessive connection-related logs from broker (#1245)
* [PINOT-4726] Fix broker errors/exceptions on connections to servers

Removed common case of server restart and flagged those as INFO messages.
If a request times out, or we cannot get a connection, then we get warnings (metrics are bumped).

Fixed a bug when an idle connection is removed by the broker. The flow that was happening was:
  AsyncPoolImpl detects an idle object and calls _lifecycle.destroy() which goes to AsyncPoolResourceManagerAdapter
  AsyncPoolResourceManagerAdapter.destroy() calls NettyClientConnection.destroy()
  NettyClientConnection closes the channel.
  As a result, we get a channelInactive callback from Netty in NettyTCPClientConnection.NettyClientConnectionHandler
  we try to destroy the 'resource' again.

* Changed class name and some messages to warn from info

* PooledNettyClientResourceManager should be managing PooledClientConnection so that we don't have to
type-cast when the pool is trying to destroy a 'resource'. NettyClientConnection has no notion of
'destroy'

* Addressed review comments