Derive partition ranges during indexing instead of relying on config. (#1258)
1. Compute the partition range during segment generation instead of
getting it from segment generation config. We will generate only one
range currently, instead of a list of ranges. Metadata and pruner still
support a list of partition ranges.

2. The new code will now always write out a partition range as computed
while indexing. This is different from the existing code that would not
write the partition range metadata if it is not as expected.

3. Updated tests to reflect this change.