Enhance IndexingConfig to have SegmentPartitionConfig. (#1264)
The current indexing config has a map of column to partition function
name, which is not sufficient information. Also, it imposes the burden
on client of IndexingConfig to parse this info and build
SegmentPartitionConfig. This change embeds SegmentPartitionConfig inside
of IndexingConfig, thus solving both the problems.

1. Added partition config inside of indexing config.
2. Separated out number of partitions as a first class member of
partition config, as opposed to embeded in partition function string.
3. Moved the config classes to config, as IndexingConfig resides there.
4. Enhanced unit tests to cover the changes.