Bugfix for server instance shutdown state (#1288)
Fix the bug where marking an instance as shutting down would correctly
remove it from the routing table, but marking it as not shutting down
would not add it back to the routing table if there was more than one
server node in the cluster. Add integration test that catches this
particular edge case. Use hostname instead of host address in
integration tests.