Fix != in query predicates (#1289)
Fix != in query predicates. The previous grammar did not support
C-style inequalities. For some reason, antlr4 is configured by default
to not reject invalid tokens when lexing the input string (!) and
simply logs the offending tokens to the console. Since ! is not a
valid token, a query that had a clause with a C-style inequality (eg.
WHERE foo != 1234) got silently converted to an equality query (eg.
WHERE foo = 1234).

This commit adds support for C-style inequality (!=) operator, as it is
supported by other databases.

Changes:
 - Add a lexer error listener that correctly throws an exception if
   there is an unexpected token in the lexing phase.
 - Add support for the != operator
 - Add support for better error messages (eg. character at which
   lexing/parsing failed)
 - Add a unit test for correct parsing of the != operator
 - Add a unit test that queries with invalid lexer tokens are correctly
   rejected
 - Add a unit test that queries with invalid parse trees are correctly
   rejected
 - Add a unit test to ensure that parse failures have proper line and
   character information
 - Remove antlr4 logging to System.err