Eliminate exception due to json parse error of table config zn record (#1325)
* Eliminate exception due to json parse error of table config zn record

If a table config zn record fails to parse json, it is possible that controller fails to
take over leadership. We scan all table config records in PinotRealtimeSegmentManager.refreshWatchers()
method. One of the ZN records in the list will be null, and will throw an NPE as we iterate over
the list of tables. We do catch the exception, but then attempt to log the table name from
znRecord.getId(), and the log message throws an NPE

* Log the NPE