Pick the first server to report a row-limit reason as winner (#1369)
* Pick the first server to report a row-limit reason as winner

In LLC, if a server reaches row limit and stops consumption, we can make a reasonable assumption
that other servers will also reach row limit. If the first server (that starts the FSM) reports
a row-limit as reason for stopping, we pick that server as the winner and get the commit protocol
going right away.

We do not pay attention to the stop reason once we have picked a winner.

Added unit tests on controller and server side.

This change can be rolled out either controller first or server first. Until both are rolled out,
the optimization will not take effect.

Fixed a typo in PARTIAL_CONSUMING state name

* Addressed review comments