Prevent checking for the field 'replication' while adding LLC tables.
LLC tables do not pay attention to repliction, only to replicasPerPartition.
In case of HLC, replication must equal to (or be a multiple of) the number of
realtime instances, checking it in PinotTableRestletResource.java and setting it to
some arbitrary configured value does not help.

For offline tables, the check is correct but should be done conditionally (i.e. table type is
OFFLINE)