Update the highest kafka offset metric correctly so that we can use it for alerting
We currently have the capability to raise an alert if the metric kafka number of rows consumed goes to 0
in order to indicate a fault in the system that can cause us to stop consuming from Kafka.
However, this metric does not give the correct picture when there are no events in the kafka
pipeline, or when the kafka pipeline is burstly (e.g. pipelines from Samza jobs).

The idea here is to use the Kafka Highest Offset Consumed metric to detect these conditions.

In order for us to use this metric effectively, we need to make sure that this metric is set to 0
when we are not consuming any rows at all.