Avoid creating unnecessary garbage during realtime indexing.
1. Instead of creating a new ArrayList for dimension and metric names,
caches those, as they are requested for each row being indexed.

2. Avoid creating MD5 hash of the schema for each row being indexed.
Since schema changes are not that frequent, it is OK to have a linear
search in small sized list, as opposed to creating a map with MD5 string
has the key.

When consuming at several thousand events per second, these could produce
few GB's of garbage within several seconds. This change eliminates
that garbage generation.