[PINOT-4885] Change FixedByteSingleColumnMultiValueReaderWriter to expand header section dynamically
Changed FixedByteSingleColumnMultiValueReaderWriter so that the header sections also expands like the data
section, so that we don't have to allocate data for max number of rows in the beginning.

For now, the same allocation is maintained so that we don't see any change in the behavior in production.

This, and the next few PRs should help us limit consumption by memory (instead of by number of rows) or time,
making it easier to configure real-time service in pinot.