[PINOT-4887] Added support for no-dictionary columns in Realtime consuming segment
Queries work on no dictionary columns in the consuming segment, but segment generation
is broken because of assumption of presense of a dictionary while getting row values.

Limiting the hanges in this PR to make it smaller. Will submit another PR to get
the segment generation right, and then enable the feature.