[PINOT-4887] Ensure that we put in an invalid value for cardinality in metadata, and throw exception when invalid value is present.
When realtime segment has no-dictionary columns, we do not compute the cardinality in RealtimeNoDictionaryColStatistics.
Instead, we return Integer.MIN_VALUE. This value gets written to the segment metadata when the segment is sealed.

When loading the metadata, we want to ensure that the bad number is not used for any processing, so we throw an
exception when it is accessed