[PINOT-4602] Changed servers to re-use segment files if available, when asked to re-commit an LLC segment
It can happen that a servers overshoot the lease (due to GC, for example) and we end up with the controller
asking a server to commit the segment again. In that case, if the offset is still the same, we keep the tar
file around and re-use it when committing the segment.

We take care to remove the tar file when we go ONLINE for a segment, so that old files are cleaned up.