Add PQL support for query options (#1490)
Add support for specifying query options through a clause in the query, similar
to the MSSQL syntax.  For example, it is possible to say SELECT * FROM foo
OPTION (bar=baz), which results in the query option bar to be assigned the
value baz.

This can eventually be used to specify query options such as partition
assignments or query optimization flags (eg. SELECT ... OPTION(partition=2,
useStarTree=true).