[PINOT-4889] Off-heap mutable dictionary for fixed width columns
Forward mapping of index to value is maintained using the FixedWidth reader-writers.
The reverse mapping is maintained using IntBuffer(s). The dictionary ID is stored as an integer
in the IntBuffer at an offset that is derived by hashing the actual value.

Collisions are handled by creating new IntBuffers, potentially larger than the previous one created.
A small overflow is maintained in heap so that we don't create new buffers on a single collision.

The current perf suite on dictionary performance is not correct for micro-benchmark.
However, a few brif runs and eye-balling results seem to show:
* Overall performance of off-heap implementation does not degrade with cardinality
 (probably because we pre-allocate the table for that cardinality in the reverse map).
* At high cardinality (> 2M) indexing in off-heap is faster (3x?).
* Value fetching time is slower in off-heap (4x?)
* Dictionary id fetching time is a little slower in off-heap for small cardinalities (< 2M) but
  faster for higher cardinality.

That said, we need to have a better perf measuring program for comparisons.

TODO:
    Limit the number of buffers created (and close the segment), so that we minimize performance
    impact on lookup. This will be done in subsequent PRs when the offheap implementation is weaved into
    RealtimeSegmentImpl

    Do some performance studies and set num-columns, overflow-threshold, expansion-multiples etc.