Fix the LLC_PARTITION_CONSUMING metric
The idea was to update this metric every 3 minutes if there is no event from Kafka.
Since we also have a timeout set when we poll kafka, the total idle time is the timeout
plus the amount of time we wait if we receive no event from kafka.

Without accounting for the kafka fetch timeout, the metric was getting updated much less
frequently, causing alerts in cases where kafka event rate is very small.