Add a REST API to fetch crc information (#1547)
In pinot-custom-metrics, crc checker is very slow and puts a lot of load to ZK because it needs
to fetch segment metadata one by one. To fix that, the rest api for fetching crc information
of a table is added. This api caches the segment crc information and its metadata version.
It only updates the cache when the segment is new or its metadata is updated to newer version.