Re-factored RetentionManager to use PinotHelixManager alone, and converted RetentionManager to a pure unit test
Re-factored RetentionManager to use PinotHelixManager exlusively for reading or writing helix metadata.

Converted RetentionManagerTest into a real unit test that does not start up a cluster or zk servers to
verify that segments are deleted.

Instead, we verify that the right call is made to PinotHelixResourceManager.

The actual deletion logic is verified in ResourceManagerTest